Add files via upload
WHAT IS FRAUD(SPAM) DETECTION ?
Fraud detection in the context of spam emails involves 	identifying and filtering out emails that are not only unsolicited but also potentially harmful or deceptive. This process is crucial for protecting users from phishing attacks, malware, and other fraudulent activities.

Fraud detection in spam emails refers to the identification and filtering of emails that are designed to deceive recipients into providing sensitive information, downloading malware, or engaging in other harmful activities. This is important for maintaining the security and integrity of email communication systems.
IMPACT OF SPAM EMAILS ON BUSINESSES
- Productivity Loss
- Financial Costs
- Security Risks
- System Performance
- Reputational Damage
- Missed Opportunities
- Resource Allocation
- Legal and Regulatory Issues
- Customer Trust
- Operational Disruption